app/requirements.txt
prometheus_client==0.20.0
PyYAML==6.0.1
vartastorage==0.2.7